Plotting functions for heidi models

plot_vs, plot_acts, plot_rs, and plot_as plot weights, activations, r-values and alphas from a model. make_plots is a convenience function to generate all the above plots in one go. plot_common_scale rescales the y-axis of a group of plots, so they are all in the same scale. get_plot_opts returns plotting options. graph_weights creates a graph of a model's weights on a given trial. get_graph_opts returns graphing options. make_graphs is a convenience function to create group graphs, given a model. patch_plots and patch_graphs return a composite of plots or graphs.

Usage

plot_vs(vals)

plot_acts(vals, bars = F)

plot_rs(vals, simple = F)

plot_as(vals)

make_plots(parsed_model)

plot_common_scale(plots)

get_plot_opts(common_scale = TRUE)

graph_weights(
  weights,
  limits = max(abs(range(weights$value))) * c(-1, 1),
  t = max(weights$trial),
  graph_opts = get_graph_opts()
)

get_graph_opts(graph_size = "small")

make_graphs(
  parsed_model,
  limits = max(abs(range(parsed_model$vs$value))) * c(-1, 1),
  t = max(parsed_model$vs$trial),
  graph_opts = get_graph_opts()
)

patch_graphs(graphs, selection = names(graphs))

patch_plots(plots, selection, plot_options = get_plot_opts())

Arguments

vals

A data.frame containing parsed values.

bars

A logical stipulating whether to summarize and use stacked bars, instead of points and lines.

simple

A logical stipulating whether to simplify the plot by collapsing across sources.

parsed_model

A parsed model, as returned by parse_heidi_results.

plots

A named list with plots

common_scale

A logical. Whether to plot the data in a common y-scale.

weights

A data.frame containing parsed weights, as returned by parse_heidi_results

limits

A vector of length 2 specifying the range of weights. Defaults to the negative and positive maximum of absolute weights.

t

An integer denoting the trial of the weights to be graphed. Defaults to the last trial in the data.

graph_opts

A list of options for graphing weights

graph_size

A string specifying the desired graph size, from c("large", "small"). Default is "large".

graphs

A list of graphs, as returned by make_graphs

selection

A character vector with the selected plots

plot_options

A list with options

See also