Parses a string stipulating trials

Parses a string stipulating trials

Usage

trial_parser(trial_string)

Arguments

trial_string

A string

Value

A list with

trial_names: A character vector specifying the names of the trials contained in the string

trial_repeats: A numeric vector specifying the number of times each trial is given

trial_pre_functional: A list of character vectors specifying the unique functional stimuli on the expectation part of each trial

trial_post_functional: A list of character vectors specifying the unique functional stimuli on the correction part of each trial

trial_pre_nominal: A list of character vectors specifying the nominal stimuli on the expectation part of each trial

trial_post_nominal: A list of character vectors specifying the nominal stimuli on the expectation part of each trial

nomi_func_map: A data.frame specifying the mapping between functional and nominal stimuli

unique_nominal stimuli: A character vector with the unique nominal stimuli

unique_functional_stimuli: A character vector with the unique functional stimuli

is_test: A logical vector specifying whether each trial should not result in learning

Note

Supports complex stimuli in parentheses (e.g., "5A>(US)") Supports multiple versions of stimuli (e.g., "3(A_a)>(US_a)/3(A_b)>(US_b)") Supports probe trials, followed by a hash (e.g., "3X#") On trials that do not have the ">" character, it is assumed that all the specified stimuli are to be used in the generation of the expectation/responding. If so, the "post" entries in the list will contain NAs, as necessary. On those trials, the "post" stimuli will be none (a character(0) vector).

Examples

trial_parser("30A")
#> $trial_names
#> [1] "A"
#> 
#> $trial_repeats
#> [1] 30
#> 
#> $trial_pre_functional
#> $trial_pre_functional[[1]]
#> [1] "A"
#> 
#> 
#> $trial_post_functional
#> $trial_post_functional[[1]]
#> character(0)
#> 
#> 
#> $trial_pre_nominal
#> $trial_pre_nominal[[1]]
#> [1] "A"
#> 
#> 
#> $trial_post_nominal
#> $trial_post_nominal[[1]]
#> character(0)
#> 
#> 
#> $nomi_func_map
#>   nomi func
#> 1    A    A
#> 
#> $unique_nominal_stimuli
#> [1] "A"
#> 
#> $unique_functional_stimuli
#> [1] "A"
#> 
#> $is_test
#> [1] FALSE
#> 
trial_parser("30XA>U/30(YA)>A")
#> $trial_names
#> [1] "XA>U"   "(YA)>A"
#> 
#> $trial_repeats
#> [1] 30 30
#> 
#> $trial_pre_functional
#> $trial_pre_functional[[1]]
#> [1] "X" "A"
#> 
#> $trial_pre_functional[[2]]
#> [1] "YA"
#> 
#> 
#> $trial_post_functional
#> $trial_post_functional[[1]]
#> [1] "U"
#> 
#> $trial_post_functional[[2]]
#> [1] "A"
#> 
#> 
#> $trial_pre_nominal
#> $trial_pre_nominal[[1]]
#> [1] "X" "A"
#> 
#> $trial_pre_nominal[[2]]
#> [1] "YA"
#> 
#> 
#> $trial_post_nominal
#> $trial_post_nominal[[1]]
#> [1] "U"
#> 
#> $trial_post_nominal[[2]]
#> [1] "A"
#> 
#> 
#> $nomi_func_map
#>   nomi func
#> 1    X    X
#> 2    A    A
#> 4   YA   YA
#> 5    U    U
#> 
#> $unique_nominal_stimuli
#> [1] "X"  "A"  "U"  "YA"
#> 
#> $unique_functional_stimuli
#> [1] "X"  "A"  "YA" "U" 
#> 
#> $is_test
#> [1] FALSE FALSE
#> 
trial_parser("1X#/2X>(US)#")
#> $trial_names
#> [1] "X#"      "X>(US)#"
#> 
#> $trial_repeats
#> [1] 1 2
#> 
#> $trial_pre_functional
#> $trial_pre_functional[[1]]
#> [1] "X"
#> 
#> $trial_pre_functional[[2]]
#> [1] "X"
#> 
#> 
#> $trial_post_functional
#> $trial_post_functional[[1]]
#> character(0)
#> 
#> $trial_post_functional[[2]]
#> [1] "US"
#> 
#> 
#> $trial_pre_nominal
#> $trial_pre_nominal[[1]]
#> [1] "X"
#> 
#> $trial_pre_nominal[[2]]
#> [1] "X"
#> 
#> 
#> $trial_post_nominal
#> $trial_post_nominal[[1]]
#> character(0)
#> 
#> $trial_post_nominal[[2]]
#> [1] "US"
#> 
#> 
#> $nomi_func_map
#>   nomi func
#> 1    X    X
#> 6   US   US
#> 
#> $unique_nominal_stimuli
#> [1] "X"  "US"
#> 
#> $unique_functional_stimuli
#> [1] "X"  "US"
#> 
#> $is_test
#> [1] TRUE TRUE
#> 
trial_parser("1X>(US_a)/1Y>(US_b)")
#> $trial_names
#> [1] "X>(US_a)" "Y>(US_b)"
#> 
#> $trial_repeats
#> [1] 1 1
#> 
#> $trial_pre_functional
#> $trial_pre_functional[[1]]
#> [1] "X"
#> 
#> $trial_pre_functional[[2]]
#> [1] "Y"
#> 
#> 
#> $trial_post_functional
#> $trial_post_functional[[1]]
#> [1] "US"
#> 
#> $trial_post_functional[[2]]
#> [1] "US"
#> 
#> 
#> $trial_pre_nominal
#> $trial_pre_nominal[[1]]
#> [1] "X"
#> 
#> $trial_pre_nominal[[2]]
#> [1] "Y"
#> 
#> 
#> $trial_post_nominal
#> $trial_post_nominal[[1]]
#> [1] "US_a"
#> 
#> $trial_post_nominal[[2]]
#> [1] "US_b"
#> 
#> 
#> $nomi_func_map
#>   nomi func
#> 1    X    X
#> 2    Y    Y
#> 4 US_a   US
#> 6 US_b   US
#> 
#> $unique_nominal_stimuli
#> [1] "X"    "Y"    "US_a" "US_b"
#> 
#> $unique_functional_stimuli
#> [1] "X"  "Y"  "US"
#> 
#> $is_test
#> [1] FALSE FALSE
#>