Plotting functions for heidi models

Plotting functions for heidi models

Usage

make_plots(dat)

plot_ws(vals)

plot_vs(vals, bars = F)

plot_rs(vals, simple = F)

plot_as(vals)

patch_plots(plots, selection, options = get_plot_opts())

get_plot_opts(common_scale = TRUE)

plot_common_scale(plots)

graph_weights(
  ws,
  limits = max(abs(range(ws$value))) * c(-1, 1),
  t = max(ws$trial),
  opts = get_graph_opts()
)

get_graph_opts(
  arrow.gap = 0.16,
  arrow.curvature = 0.2,
  arrow.pt = 20,
  edge.size = 3,
  node.size = 40,
  node.stroke = 3,
  node.text.size = 15
)

make_graphs(
  mod,
  limits = max(abs(range(mod$ws$value))) * c(-1, 1),
  trial = max(mod$ws$trial)
)

patch_graphs(graphs)

Arguments

dat

A list containing data.frames with parsed weights, r values, and v values, as returned by parse_heidi_results.

vals

A data.frame containing parsed values

bars

A logical stipulating whether to summarize and use stacked bars, instead of points and lines.

simple

A logical stipulating whether to simplify the plot by collapsing across sources.

plots

A named list with plots

selection

A character vector with the selected plots

options

A list with options

common_scale

A logical. Whether to plot the data in a common y-scale.

ws

A data.frame containing parsed weights, as returned by parse_heidi_results

limits

A vector of length 2 specifying the range of weights. Defaults to the negative and positive maximum of absolute weights.

t

An integer specifying a trial

opts

A list of options for graphing weights. See ?get_graph_opts.

arrow.gap

Distance between nodes and arrows

arrow.curvature

Curvature of the arrows

arrow.pt

Size of the arrows (see grid::arrow)

edge.size

Thickness of the arrows

node.size

Size of the nodes

node.stroke

Thickness of the node lines

node.text.size

Thickness of the text within the node

mod

Parsed model results, as returned by parse_heidi_results.

trial

An integer denoting the trial of the weights to be graphed. Defaults to the last trial in the data.

graphs

A list of graphs, as returned by make_graphs

Value

A list with ggplot objects

Note

Plotting options are obtained via get_plot_opts(). For now, only plotting in a common y-axis is supported.

See also